Add files via upload
It is a Python gui made my me using tkinter class in python. It works a an online restaurant system. it gives the user choice different types of food variety and make him choose accordingly what he wants and how much quantity of it he wants and then show all the order in form of menu and thus calculate the bill accordingly.